requirements.txt
asgiref==3.5.2
Django==4.1.2
django-bootstrap-datepicker-plus==5.0.0
django-bootstrap5==22.1
djangorestframework==3.14.0
psycopg2-binary==2.9.4
pydantic==1.10.2
pytz==2022.6
sqlparse==0.4.3
typing_extensions==4.4.0
